Create a header-only Souffle library, use instead of `souffle_lib`.
On chat, we had a discussion as to whether Souffle needs to be linked
into the C++ programs generated from compiled Souffle or whether they
are standalone programs. This commit is an experiment to clarify that:
in all cases possible, this replaces `souffle_lib` with
`souffle_include_lib`, a library consisting only of Souffle header
files. It appears that this can replace `souffle_lib` in all cases in
Raksha where we are compiling Datalog, meaning that Souffle-generated
C++ programs do not need to link with Souffle.

Note that we would need to link with a native binary if we were to use
Souffle user-defined functors. However, we have not needed to resort to
that yet.