Allow disabling authorization logic with compile flag. (#294)
* Allow disabling authorization logic with compile flag.

Refactor the bridge between the Rust authorization logic compiler and
the C++ xform_to_datalog tool to allow cutting the connection by
defining `DISABLE_AUTHORIZATION_LOGIC`. To encapsulate this macro within
a single package, refactor the bridge to place the `extern "C"`
declaration `generate_datalog_facts_from_authorization_logic` into a .cc
file and create a C++ bridge method to encapsulate it.

* Move comment closer to code it explains.

* Make the failure when running with auth logic disabled friendlier.

Previously, when `DISABLE_AUTHORIZATION_LOGIC` was defined, we would just
have `GenerateDatalogFactsFromAuthorizationLogic` just return 1. This
would cause a failure, but it had the potential to be cryptic. Refactor
to use a `CHECK` to provide an error message, reducing potential for
confusion.

* Add comment explaining compile-time disabling of auth logic.